% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb.density.R
\name{nb.density}
\alias{nb.density}
\alias{ldnb}
\alias{dnb}
\title{dnb, ldnb Functions}
\usage{
ldnb(x, theta, mu)

dnb(x, theta, mu)
}
\arguments{
\item{x}{A positive numeric scalor or vector. Decimals and integers are both allowed.}

\item{theta}{Value of dispersion.}

\item{mu}{Value of mean.}
}
\value{
\item{dnb}{Density of generalized Negative Binomial} 
       \item{ldnb}{Log-density of generalized Negative Binomial}
}
\description{
These functions allow you to compute (log-)density of generalized Negative Binomial distribution.
}
\examples{
ldnb(x=10.4,theta=3.2,mu=5)
dnb(x=10.4,theta=3.2,mu=5)
}
