% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulated_Count.R
\docType{data}
\name{Simulated_Count}
\alias{Simulated_Count}
\title{Data set for illustration: Simulated_Count}
\format{A simulated data frame with 50 rows (i.e. samples) and 100 columns (i.e. variables ).
        It can be viewed as simulated RNA-Seq integer counts of 100 genes for 50 patients.}
\usage{
Simulated_Count
}
\description{
Data set for illustration: Simulated_Count
}
\keyword{datasets}
